Issue #116 refactor: controller 제거
